Fix bug with losing count and refactor
